save upload file user

+++ api/controllers/BusinessProcessController.coffee
@@ -4,10 +4,15 @@ actionUtil = require 'sails/lib/hooks/blueprints/actionUtil'
 create = require 'sails/lib/hooks/blueprints/actions/create'
 
 getDeploymentDetails = (procDef) ->
-	activiti.req 'get', sails.config.activiti.url.deployment procDef.deploymentId
-		.then (result) ->
-			procDef.deploymentDetails = result.body
-			return procDef
+	sails.models.businessprocess
+		.find({deploymentId: procDef.deploymentId})
+		.then (rst) ->
+			activiti.req 'get', sails.config.activiti.url.deployment procDef.deploymentId
+				.then (result) ->
+					procDef.deploymentDetails = result.body
+					if rst.length > 0
+						procDef.createdBy = rst[0].createdBy
+					return procDef
 		.catch (err) ->
 			sails.log.error err
 						
@@ -21,24 +26,23 @@ module.exports =
 		req.file('file').upload fileOpts, (err, files) ->
 			if err
 				return res.serverError(err)
-
 			data = 
 				file: { file: files[0].fd, content_type: 'multipart/form-data' }
-			
 			activiti.deployXML "#{sails.config.activiti.url.deployment ''}", data
 				.then (rst) ->
-					sails.log.info "Del file: #{files[0].fd}"
-					fs.unlinkSync "#{files[0].fd}"
-							
-					data =
-						deploymentId:	rst.body.id
-						deploymentTime:	rst.body.deploymentTime
-						filename:		rst.body.name
-						createdBy:		req.user.username
-					sails.models.businessprocess.create(data)
-						.then (newInstance) ->
-							res.ok newInstance
-
+					if rst.statusCode == 201  
+						fs.unlinkSync "#{files[0].fd}"
+						data =
+							deploymentId:	rst.body.id
+							deploymentTime:	rst.body.deploymentTime
+							filename:		rst.body.name
+							createdBy:		req.user.username
+						sails.models.businessprocess.create(data)
+							.then (newInstance) ->
+								res.ok newInstance
+					else
+						sails.log.error JSON.stringify rst.body
+						return res.serverError(rst.body)
 				.catch res.serverError
 				
 	find: (req, res) ->
@@ -59,6 +63,6 @@ module.exports =
 			.then (processdefList) ->
 				result = _.findWhere(processdefList.body,{type: 'processDefinition'})
 				activiti.getXML "#{sails.config.activiti.url.deployment data.deploymentId}/resourcedata/#{result.id}"
-					.then (stream) ->
-						res.ok(stream.raw)
+			.then (stream) ->
+				res.ok(stream.raw)
 			.catch res.serverError
